[move]: navigation 패키지 분리
